Refactoring and commented out stuff we do not need

+++ CsGoApplicationAimbot/CsGoApplicationAimbot/CSGOClasses/Geometry.cs
@@ -7,7 +7,7 @@ namespace CsGoApplicationAimbot.CSGOClasses
     {
         public static float GetDistanceToPoint(Vector3 pointA, Vector3 pointB)
         {
-            return (float)Math.Abs((pointA - pointB).Length());
+            return Math.Abs((pointA - pointB).Length());
         }
     }
 }
